ปรับ rdu

+++ surasak3/rdu_convert.php
@@ -22,20 +22,31 @@ function dump($txt){
     echo "</pre>";
 }
 
-include 'includes/connect_sv13.php';
+// include 'includes/connect_sv13.php';
 
 // mysql_query('SET NAMES TIS620', $db);
+define('HOST', '192.168.131.250');
+define('PORT', '3306');
+define('DB', 'smdb');
+define('USER', 'remoteuser');
+define('PASS', '');
 
-$date_start = '2564-07-01';
-$date_end = '2564-09-15';
+$dbi = new mysqli(HOST,USER,PASS,DB);
 
-$quarter = 4;
-$year = '2564';
+
+$date_start = '2565-04-01';
+$date_end = '2565-06-30';
+
+$quarter = 3;
+$year = '2565';
 
 $dirPath = realpath(dirname(__FILE__))."/rdu";
 $filePath = $dirPath.'/'.$date_start.'_'.$date_end.'_opday_'.$quarter.'.sql';
+if(file_exists($filePath))
+{
+    unlink($filePath);
+}
 
-unlink($filePath);
 
 $sql = "SELECT a.`row_id`,a.`thidate`,a.`hn`,a.`ptname`,a.`age`,a.`diag`,a.`icd10`,a.`doctor`,SUBSTRING(TRIM(a.`toborow`),1,4) AS `toborow`, 
 CONCAT(SUBSTRING(a.`thidate`,1,10),a.`hn`) AS `date_hn` 
@@ -44,15 +55,16 @@ WHERE a.`thidate` >= '$date_start 00:00:00' AND a.`thidate` <= '$date_end 23:59:
 AND a.`an` IS NULL 
 AND ( a.`icd10` <> '' AND a.`icd10` IS NOT NULL ) 
 AND a.`doctor` <> '' ";
-$q = mysql_query($sql, $db) or die( mysql_error() );
-
+// $q = mysql_query($sql, $db) or die( mysql_error() );
+$q = $dbi->query($sql);
 
 $sql_header = "INSERT INTO `opday` ( `id`,`row_id`,`date`,`hn`,`ptname`,`gender`,`age`,`diag`,`icd10`,`doctor`,`toborow`,`date_hn`,`date_generate`,`quarter`,`year`) VALUES ";
 $sql_data_list = '';
 
 $test_i = 0;
 
-while ( $item = mysql_fetch_assoc($q) ) {
+// while ( $item = mysql_fetch_assoc($q) ) {
+while ( $item = $q->fetch_assoc() ) {
 
     $row_id = $item['row_id'];
     $thidate = $item['thidate'];
@@ -84,6 +96,6 @@ while ( $item = mysql_fetch_assoc($q) ) {
     $test_i++;
 }
 
-mysql_close($db);
+// mysql_close($db);
 
 echo "Success";
