+++ surasak3/rdu_convert_drug.php
@@ -8,19 +8,29 @@ function dump($txt){
     echo "</pre>";
 }
 
-include 'includes/connect_sv13.php';
+// include 'includes/connect_sv13.php';
 
 // mysql_query('SET NAMES TIS620', $db);
 
-$date_start = '2564-06-01';
-$date_end = '2564-06-31';
+define('HOST', '192.168.131.250');
+define('PORT', '3306');
+define('DB', 'smdb');
+define('USER', 'remoteuser');
+define('PASS', '');
+
+$dbi = new mysqli(HOST,USER,PASS,DB);
+
+$date_start = '2565-04-01';
+$date_end = '2565-06-30';
 $quarter = 3;
-$year = '2564';
+$year = '2565';
 
 $dirPath = realpath(dirname(__FILE__))."/rdu";
 $filePath = $dirPath.'/'.$date_start.'_'.$date_end.'_drug_'.$quarter.'.sql';
-
-unlink($filePath);
+if(file_exists($filePath))
+{
+    unlink($filePath);
+}
 
 $sql = "SELECT `row_id`,`date`,`hn`,`drugcode`,TRIM(`part`) AS `part`,`amount`,CONCAT(SUBSTRING(`date`,1,10),`hn`) AS `date_hn` 
 FROM `drugrx` 
@@ -28,13 +38,14 @@ WHERE `date` >= '$date_start 00:00:00' AND `date` <= '$date_end 23:59:59'
 AND `an` IS NULL 
 AND `reject` != 'Y' 
 AND `amount` > 0 ";
-$q = mysql_query($sql, $db) or die( mysql_error() );
+// $q = mysql_query($sql, $db) or die( mysql_error() );
+$q = $dbi->query($sql);
 
 $sql_header = "INSERT INTO `drugrx` ( `id`,`row_id`,`date`,`hn`,`drugcode`,`part`,`amount`,`date_hn`,`date_generate`,`quarter`,`year`) VALUES ";
 $sql_data = '';
 
-while ( $item = mysql_fetch_assoc($q) ) {
-
+// while ( $item = mysql_fetch_assoc($q) ) {
+while ( $item = $q->fetch_assoc() ) {
     $row_id = $item['row_id'];
     $date = $item['date'];
     $hn = $item['hn'];
@@ -48,6 +59,6 @@ while ( $item = mysql_fetch_assoc($q) ) {
 
 }
 
-mysql_close($db);
+// mysql_close($db);
 
 echo "Success";
